Move test configs to `test-utils`
