+++ apps/console/test-utils/redux-store-state-mock.json
